Fix the mnist test for keras quantization api
PiperOrigin-RevId: 282624959

+++ tensorflow_model_optimization/python/core/quantization/keras/utils.py
@@ -32,7 +32,7 @@ def convert_keras_to_tflite(model_path,
 
   if is_quantized:
     converter.inference_type = tf.lite.constants.INT8
-    converter.inference_input_type = tf.lite.constants.INT8
+    converter.inference_input_type = tf.lite.constants.FLOAT
 
     input_arrays = converter.get_input_arrays()
     converter.quantized_input_stats = {
